move io function to IOcallback
